Move unit tests into the `unit` folder
